Scale seed: 1,600 fake deals + fixes the big-book surfaces
- scripts/seed-scale.ts generates a deterministic full book (~1,600 converted
  deals, ~780 customers, ~40k ledger transactions, arrears + collections
  workflows) into the local DB and writes additive supabase-scale-deals-*.sql
  files (explicit ids above 100000) for loading the same book into the live DB
- DataTable renders at most 250 rows (filter/sort/CSV still cover everything):
  asset register went from 2.7s / 1.9MB to 0.4s / 0.9MB
- Aggregate sums cast to float8 — the portfolio total overflowed int4 at $221M
- Dashboard: compact $xxx.xM stat, insurance-expiring list capped at 16
- Application references now continue from the highest existing reference
  number rather than the row id
- scripts/perf.mjs: page-load timing harness

+++ src/app/(app)/accounts/page.tsx
@@ -19,8 +19,8 @@ export default async function LoansPage() {
     const sums = await db
       .select({
         loanId: transactions.loanId,
-        // sum() over integers is bigint in Postgres — cast so the driver returns a number
-        balance: sql<number>`coalesce(sum(${transactions.amountExGstCents}), 0)::int`,
+        // sum() over integers is bigint in Postgres — cast to float8 so the driver returns a number without overflowing int4 on a large book
+        balance: sql<number>`coalesce(sum(${transactions.amountExGstCents}), 0)::float8`,
       })
       .from(transactions)
       .where(inArray(transactions.loanId, loanIds))
